requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.1
pydantic==2.9.2
pyyaml==6.0.2
httpx==0.28.1
python-multipart==0.0.17
loguru==0.7.3 